clean up package.py, rename spotmessanger.txt to customize.txt

+++ package.py
@@ -5,17 +5,14 @@ import sys
 import shutil
 import argparse
 import re
+import ConfigParser
 from string import Template
 
-WOT_VERSION          = "0.9.18 Common Test"
-SUPPORT_URL          = ""
-ROOT_DIR             = os.path.dirname(os.path.realpath(__file__))
-SRC_DIR              = os.path.join(ROOT_DIR, "src")
-SCRIPT_DIR           = os.path.join(ROOT_DIR, "src", "scripts")
-CONFIG_DIR           = os.path.join(ROOT_DIR, "configs")
-BUILD_DIR            = os.path.join(ROOT_DIR, "build")
-
-MOD_BASE_DIR         = os.path.join(SRC_DIR, "scripts", "client", "gui", "mods")
+ROOT_DIR            = os.path.dirname(os.path.realpath(__file__))
+SCRIPT_DIR          = os.path.join(ROOT_DIR, "src", "scripts")
+CONFIG_DIR          = os.path.join(ROOT_DIR, "configs")
+BUILD_DIR           = os.path.join(ROOT_DIR, "build")
+SCRIPT_DEBUG_BASE   = "scripts"
 
 DOC_FILES = [
     os.path.join(ROOT_DIR, "LICENSE"),
@@ -27,26 +24,33 @@ META_FILES = [
 ]
 
 sys.dont_write_bytecode = True
-sys.path.append(MOD_BASE_DIR)
-from spotmessanger.version import MOD_INFO
 
 def main():
+    inifile = ConfigParser.SafeConfigParser()
+    inifile.read('config.ini')
+    wot_version = inifile.get('wot', 'version')
+
     parser = argparse.ArgumentParser()
-    parser.add_argument('--mod-debug',   default=MOD_INFO.DEBUG)
+    parser.add_argument('-d', '--debug', action='store_true', help='show debug messages')
+    parser.add_argument('-k', '--keep-tmpfile', action='store_true', help='keep tempfile')
+    parser.add_argument('--mod-debug', default=inifile.getboolean('mod', 'debug'),
+            metavar='DEBUG', help='specify mod debug mode, True/False')
     args = parser.parse_args()
+    
     in_file_parameters = dict(
-        SUPPORT_URL = SUPPORT_URL,
-        DEBUG       = args.mod_debug,
-        MOD_ID      = MOD_INFO.ID,
-        MOD_NAME    = MOD_INFO.NAME,
-        MOD_VERSION = MOD_INFO.VERSION
+        DEBUG               = args.mod_debug,
+        MOD_ID              = inifile.get('mod', 'id'),
+        MOD_NAME            = inifile.get('mod', 'name'),
+        MOD_VERSION         = inifile.get('mod', 'version'),
+        MOD_VERSION_LONG    = inifile.get('mod', 'version_long'),
+        DESCRIPTION         = inifile.get('mod', 'description').replace('\n', ' '),
+        SUPPORT_URL         = inifile.get('mod', 'support_url')
     )
 
     packager = Packager(
-        src_dir               = SRC_DIR,
-        build_dir             = BUILD_DIR,
-        in_file_parameters    = in_file_parameters,
-        ignored_file_patterns = [".+_test.py"]
+        args                    = args,
+        in_file_parameters      = in_file_parameters,
+        ignored_file_patterns   = [".+_test.py"]
     )
     
     stage0 = os.path.join(BUILD_DIR, "stage0")
@@ -54,10 +58,10 @@ def main():
     stage2 = os.path.join(BUILD_DIR, "stage2")
     stage3 = BUILD_DIR
     
-    mod_name = in_file_parameters['MOD_NAME']
-    mod_version = in_file_parameters['MOD_VERSION']
-    pack_wotmod = "{name}-{version}.wotmod".format(name=mod_name, version=mod_version).lower()
-    zip_wotmod = "{name}-{version}.zip".format(name=mod_name, version=mod_version).lower()
+    mod_name = in_file_parameters['MOD_NAME'].lower()
+    mod_version = in_file_parameters['MOD_VERSION'].lower()
+    pack_wotmod = "{name}-{version}.wotmod".format(name=mod_name, version=mod_version)
+    zip_wotmod = "{name}-{version}.zip".format(name=mod_name, version=mod_version)
     
     stage = [
         [ [ stage0, "scripts"                           ], [ SCRIPT_DIR         ] ],
@@ -66,8 +70,9 @@ def main():
         [ [ stage0, "doc"                               ], [ DOC_FILES          ] ],
         [ [ stage1, "res", "scripts"                    ], [ stage0, "scripts"  ] ],
         [ [ stage1                                      ], [ stage0, "meta"     ] ],
-        [ [ stage2, "mods", WOT_VERSION, pack_wotmod    ], [ stage1             ] ],
+        [ [ stage2, "mods", wot_version, pack_wotmod    ], [ stage1             ] ],
         [ [ stage2, "mods", "configs"                   ], [ stage0, "configs"  ] ],
+        [ [ stage2, "mods", "configs", mod_name         ], [ stage0, "doc"      ] ],
         [ [ stage2                                      ], [ stage0, "doc"      ] ],
         [ [ stage3, zip_wotmod                          ], [ stage2             ] ]
     ]
@@ -88,10 +93,10 @@ def accepts_extensions(extensions):
     if passed filepath has extension in list of given 'extensions'.
     '''
     def decorator(original_function):
-        def wrapper(self, src_filepath, dest_filepath):
+        def wrapper(self, src_filepath, dest_filepath, **kwargs):
             for extension in extensions:
                 if src_filepath.lower().endswith(extension.lower()):
-                    return original_function(self, src_filepath, dest_filepath)
+                    return original_function(self, src_filepath, dest_filepath, **kwargs)
             return False
         return wrapper
     return decorator
@@ -111,8 +116,7 @@ class CallbackList(object):
 class Packager(object):
 
     def __init__(self, **dict):
-        self.__src_dir = os.path.normpath(dict['src_dir'])
-        self.__build_dir = os.path.normpath(dict['build_dir'])
+        self.__args = dict['args']
         self.__in_file_parameters = dict['in_file_parameters']
         self.__ignored_file_patterns = dict['ignored_file_patterns']
         self.__builders = CallbackList(
@@ -151,38 +155,50 @@ class Packager(object):
                     yield os.path.normpath(os.path.join(root, filename))
 
     @accepts_extensions([".py"])
-    def __compile_py_file(self, src_filepath, dest_dir):
-        '''Compiles 'src_filepath' python source file into python bytecode file and
-        saves it dest_filepath.
-        '''
-        debug_filepath = os.path.relpath(src_filepath, self.__src_dir)
+    def __compile_py_file(self, src_filepath, dest_dir, orig_filepath=None, **kwargs):
+        '''Compiles 'src_filepath' python source file into python bytecode file and saves it dest_filepath.'''
+        rel_filepath = os.path.relpath(orig_filepath or src_filepath, SCRIPT_DIR)
+        debug_filepath = os.path.join(SCRIPT_DEBUG_BASE, rel_filepath)
         dest_filepath = os.path.join(dest_dir, os.path.splitext(os.path.basename(src_filepath))[0] + ".pyc")
         self.__make_parent_dirs(dest_filepath)
         # compile source py-file into bytecode pyc-file
+        if self.__args.debug:
+            print "compile: {} -> {} ({})".format(os.path.relpath(src_filepath, ROOT_DIR),
+                    os.path.relpath(dest_filepath, ROOT_DIR), debug_filepath)
         py_compile.compile(file=src_filepath, cfile=dest_filepath, dfile=debug_filepath, doraise=True)
         return True
 
     @accepts_extensions([".swf", ".txt", ".json", ".xml", ".png", ".pyc", ".md", "LICENSE"])
-    def __copy_file(self, src_filepath, dest_dir):
+    def __copy_file(self, src_filepath, dest_dir, **kwargs):
         '''Simply copies file at 'src_filepath' to 'dest_filepath'.'''
         dest_filepath = os.path.join(dest_dir, os.path.basename(src_filepath))
         self.__make_parent_dirs(dest_filepath)
         # simply copy file from source to build dir
+        if self.__args.debug:
+            print "copy: {} -> {}".format(os.path.relpath(src_filepath, ROOT_DIR), os.path.relpath(dest_filepath, ROOT_DIR))
         shutil.copyfile(src_filepath, dest_filepath)
         return True
 
-    @accepts_extensions([".in", ".in.xml"])
-    def __run_template_file(self, src_filepath, dest_dir):
+    @accepts_extensions([".in", "in.txt", ".in.xml", ".in.py"])
+    def __run_template_file(self, src_filepath, dest_dir, **kwargs):
         parameters = self.__in_file_parameters
-        i = src_filepath.rfind('.in')
-        build_filepath = os.path.join(self.__build_dir, os.path.basename(src_filepath[:i] + src_filepath[i+3:]))
+        src_dir, src_file = os.path.split(src_filepath)
+        build_dir = os.path.join(BUILD_DIR, 'tmp')
+        i = src_file.rfind('.in')
+        filename = src_file[:i] + src_file[i+3:]
+        build_filepath = os.path.join(build_dir, filename)
+        self.__make_parent_dirs(build_filepath)
         # run 'parameters' through in-template and produce temporary output file
+        if self.__args.debug:
+            print "template: {} -> {}".format(os.path.relpath(src_filepath, ROOT_DIR), os.path.relpath(build_filepath, ROOT_DIR))
         with open(src_filepath, "r") as in_file, open(build_filepath, "w") as out_file:
             out_file.write(Template(in_file.read()).substitute(**parameters))
         # further process the output file with other builders
-        self.__builders(build_filepath, dest_dir)
+        self.__builders(build_filepath, dest_dir, orig_filepath=os.path.join(src_dir, filename))
         # remove the temporary output file
-        os.remove(build_filepath)
+        if not self.__args.keep_tmpfile:
+            os.remove(build_filepath)
+            os.rmdir(build_dir)
         return True
 
     def __make_parent_dirs(self, filepath):
